requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
pyparsing==3.2.3
sqlparse==0.5.3
tzdata==2025.2
